refactor: move AccountSetting to dedicated file
